moved test_subset to test_operations
